Change user channel subscriptions to track connections
A user channel subscription tracks the connections that have initiated the
subscription - when all connections associated with the subscription have 
been removed, the user is unsubscribed from the channel. 

The connection that first successfully subscribes the user to a channel is 
associated with the user channel subscription. Subsequent connections that
attempt to subscribe the user to the channel are also tracked (note, however,
that the webhook is not called for these subsequent connections, since it is
assumed that the first connection's success extends to them as well), so the 
first connection may be closed without unsubscribing the user, so long as a 
later connection remains open.